update shiny-react project first page, move starwars project to separate page
