+++ 1.Next-React/shiny-react/src/components/index.js
@@ -1,1 +1,2 @@
+export { default as StarWarsContent } from "./StarWarsContent";
 export { default as ToDo } from "./ToDo";
