+++ 1.Next-React/shiny-react/src/pages/StarWarsPage/index.js
@@ -1,0 +1,15 @@
+import React, { Component } from "react";
+import "./index.css";
+import { StarWarsContent } from "../../components";
+
+class StarWarsPage extends Component {
+  render() {
+    return (
+      <>
+        <StarWarsContent />
+      </>
+    );
+  }
+}
+
+export default StarWarsPage;
