+++ 1.Next-React/shiny-react/src/pages/index.js
@@ -1,2 +1,3 @@
 export { default as BalanceCheckPage } from "./BalanceCheckPage";
+export { default as StarWarsPage } from "./StarWarsPage";
 export { default as ToDoPage } from "./ToDoPage";
